+++ sdk/keyvault/azure-keyvault-administration/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Other Changes
 - Python 2.7 is no longer supported. Please use Python version 3.6 or later.
+- Updated minimum `azure-core` version to 1.20.0
 
 ## 4.1.0b2 (2021-11-11)